requirements.txt
llama-index==0.9.48
sentence-transformers
faiss-cpu
tqdm
black
ruff
requests
python-dotenv
